Add ability to include a styled map.
If the user adds a properly formatted array of styles (via new options page),
then the static map image is styled. For the dynamic map, we need to use the
JavaScript API, and therefore, an API key. So we check whether an API key is present
in the options. If it is, then we prep a full-on JS map.

Note: right now we are only doing the JS map if there are styles present.
Seems like someone might want the JS map even w/o styles.

Also: want to move more of the settings to the options page.